README.md
the process of editing film or video footage to have a new look to help make a video feel more vibrant and stylistic. Define color correcting. Color correcting is the process of editing film or video footage to fix any imbalances in the color so it can appear how the eye is meant to see it